rename and create data models for dynamic maps
